# ByteMarket/backend/src/main/java/es/grupo04/backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/v1/products/{id}:
    get:
      summary: Retrieve a product by its ID
      tags:
      - product-rest-controller
      operationId: getProductById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    put:
      summary: Update an existing product by its ID
      tags:
      - product-rest-controller
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: newProductDTO
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/NewProductDTO"
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    delete:
      summary: Delete a product by its ID
      tags:
      - product-rest-controller
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
...
  /api/v1/reports/products/{productId}:
    post:
      summary: Report a product by its ID
      tags:
      - review-report-rest-controller
      operationId: postReport
      parameters:
      - name: reportDTO
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/NewReportDTO"
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReportDTO"
  /api/v1/products:
    get:
      summary: Retrieve a paginated list of all products with optional filters
      tags:
      - product-rest-controller
      operationId: getAllProducts
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 8
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: category
        in: query
        required: false
        schema:
          type: string
      - name: recommended
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageProductDTO"
    post:
      summary: Create a new product
      tags:
      - product-rest-controller
      operationId: createProduct
      parameters:
      - name: productDTO
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/NewProductDTO"
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
  /api/v1/products/{id}/images:
    post:
      summary: Upload an image for a specific product by its ID
      tags:
      - product-rest-controller
      operationId: addImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
              required:
              - image
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
  /api/v1/products/{id}/favorites:
    post:
      summary: Toggle a product as a favorite by its ID
      tags:
      - product-rest-controller
      operationId: toggleFavoriteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 8
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageProductDTO"
...
  /api/v1/products/sales:
    get:
      summary: Retrieve a list of the user’s recent sales
      tags:
      - product-rest-controller
      operationId: getLastSales
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductDTO"
  /api/v1/products/purchases:
    get:
      summary: Retrieve a list of the user’s recent purchases
      tags:
      - product-rest-controller
      operationId: getLastPurchases
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductDTO"
  /api/v1/products/favorites:
    get:
      summary: Fetch a paginated list of the user’s favorite products
      tags:
      - product-rest-controller
      operationId: getFavoriteProducts
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 8
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageProductDTO"
...
  /api/v1/products/{productId}/images/{imageId}:
    delete:
      summary: Remove an image from a product by product and image IDs
      tags:
      - product-rest-controller
      operationId: removeImage
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: imageId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
components:
  schemas:
    NewProductDTO:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
          format: float
        category:
          type: string
        description:
          type: string
        imageUpload:
          type: array
          items:
            type: string
            format: binary
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: float
        category:
          type: string
        description:
          type: string
        owner:
          $ref: "#/components/schemas/UserBasicDTO"
        imageUrls:
          type: array
          items:
            type: string
        thumbnail:
          type: string
        sold:
          type: boolean
        publishDate:
          type: string
          format: date
    UserBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        image:
          type: string
        hasImage:
          type: boolean
...
    NewReportDTO:
      type: object
      properties:
        reason:
          type: string
        description:
          type: string
        productId:
          type: integer
          format: int64
        reportCreatorId:
          type: integer
          format: int64
    ProductBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: float
        category:
          type: string
        sold:
          type: boolean
        publishDate:
          type: string
          format: date
    ReportDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        reason:
          type: string
        description:
          type: string
        product:
          $ref: "#/components/schemas/ProductBasicDTO"
        reportCreator:
          $ref: "#/components/schemas/UserBasicDTO"
    PageProductDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProductDTO"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...