# ByteMarket/backend/src/main/java/es/grupo04/backend/controller/ProductRestController.java
package es.grupo04.backend.controller;

import java.io.IOException;
import java.net.URI;
import java.security.Principal;
import java.util.List;
import java.util.NoSuchElementException;
import java.util.Optional;
import java.util.stream.Collectors;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageImpl;
import org.springframework.data.domain.PageRequest;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.ModelAttribute;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;
import static org.springframework.web.servlet.support.ServletUriComponentsBuilder.fromCurrentRequest;

import es.grupo04.backend.dto.NewProductDTO;
import es.grupo04.backend.dto.ProductDTO;
import es.grupo04.backend.dto.UserBasicDTO;
import es.grupo04.backend.dto.UserDTO;
import es.grupo04.backend.service.ProductService;
import es.grupo04.backend.service.UserService;
import io.swagger.v3.oas.annotations.Operation;
import jakarta.servlet.http.HttpServletRequest;

@RestController
@RequestMapping("/api/v1/products")
public class ProductRestController {

    @Autowired
    private ProductService productService;

    @Autowired
    private UserService userService;

    @Operation (summary= "Retrieve a list of products")
    @GetMapping
    public Page<ProductDTO> getAllProducts(@RequestParam(defaultValue = "0") int page,
            @RequestParam(defaultValue = "8") int size, @RequestParam(required = false) String name,
            @RequestParam(required = false) String category, @RequestParam(required = false) boolean recommended){
        
        Page<ProductDTO> productsPage;
        if(recommended){
            List<ProductDTO> topRated = productService.findTopRatedSellersProducts();
            return new PageImpl<>(topRated, PageRequest.of(page, size), topRated.size());
        } else{
            if (category != null && !category.isEmpty()) {
                productsPage = productService.findAvailableByCategory(category, page, size);
            } else {
                productsPage = productService.findPaginated(PageRequest.of(page, size));
            }
            if (name != null && !name.isEmpty()) {
                List<ProductDTO> filteredByName = productsPage.getContent().stream()
                    .filter(product -> product.name().toLowerCase().contains(name.toLowerCase()))
                    .collect(Collectors.toList());
                    return new PageImpl<>(filteredByName, PageRequest.of(page, size), filteredByName.size());
            }
        }
        return productsPage;
    }
    
    @Operation (summary= "Retrieve a product by its ID")
    @GetMapping("/{id}")
    public Optional<ProductDTO> getProductById(@PathVariable Long id) {
        return productService.findById(id);
    }

    @Operation (summary= "Create a new product")
    @PostMapping
    public ResponseEntity<ProductDTO> createProduct(@ModelAttribute NewProductDTO productDTO,
            HttpServletRequest request) throws IOException {
        Principal principal = request.getUserPrincipal();
        if (principal == null) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        UserBasicDTO userDTO = userService.findByMail(principal.getName()).get();

        ProductDTO newProduct = productService.save(productDTO, userDTO.id());
        if (newProduct == null) {
            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
        }

        URI location = fromCurrentRequest().path("/{id}").buildAndExpand(newProduct.id()).toUri();
        return ResponseEntity.created(location).body(newProduct);
    }

    @Operation (summary= "Update product by its ID")
    @PutMapping("/{id}")
    public ResponseEntity<ProductDTO> updateProduct(@PathVariable Long id, @ModelAttribute NewProductDTO newProductDTO,
            HttpServletRequest request) {
        Principal principal = request.getUserPrincipal();
        if (principal == null) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        Optional<ProductDTO> productOpt = productService.findById(id);
        if (productOpt.isEmpty()) {
            throw new NoSuchElementException();
        }

        ProductDTO productDTO = productOpt.get();
        UserBasicDTO userDTO = userService.findByMail(principal.getName()).get();
        if (productDTO.owner().id() != userDTO.id()) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        Optional<ProductDTO> editedProduct = productService.updateProduct(productDTO, newProductDTO);
        if (!editedProduct.isPresent()) {
            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
        }
        return ResponseEntity.ok(editedProduct.get());
    }

    @Operation (summary= "Delete product by ID")
    @DeleteMapping("/{id}")
    public ResponseEntity<ProductDTO> deleteProduct(@PathVariable Long id, HttpServletRequest request) {
        Principal principal = request.getUserPrincipal();
        if (principal == null) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        Optional<ProductDTO> productOpt = productService.findById(id);
        if (productOpt.isEmpty()) {
            throw new NoSuchElementException();
        }

        ProductDTO productDTO = productOpt.get();
        UserBasicDTO userDTO = userService.findByMail(principal.getName()).get();
        if (productDTO.owner().id() != userDTO.id()) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        productService.delete(id);
        return ResponseEntity.ok(productDTO);
    }

    @Operation (summary= "Retrieve a list of favorites products of a user")
    @GetMapping("/favorites")
    public ResponseEntity<Page<ProductDTO>> getFavoriteProducts(
            HttpServletRequest request,
            @RequestParam(defaultValue = "0") int page,
            @RequestParam(defaultValue = "8") int size) {

        Principal principal = request.getUserPrincipal();
        if (principal == null) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        UserDTO userDTO = userService.findByMailExtendedInfo(principal.getName()).get();

        Page<ProductDTO> productsPage = productService.getFavoriteProducts(userDTO, page, size);
        return ResponseEntity.ok(productsPage);
    }
    
    @Operation (summary= "Toggle a product as a favorite by its ID")
    @PostMapping("/{id}/favorites")
    public ResponseEntity<Page<ProductDTO>> toggleFavoriteProduct(HttpServletRequest request,
            @PathVariable Long id,
            @RequestParam(defaultValue = "0") int page,
            @RequestParam(defaultValue = "8") int size) {
        
        Principal principal = request.getUserPrincipal();
        if (principal == null) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }
        if (productService.findById(id).isEmpty()) {
            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
        }
    
        ProductDTO productDTO = productService.findById(id).get();
        UserDTO userDTO = userService.findByMailExtendedInfo(principal.getName()).get();
    
        if (productDTO.owner().id() == userDTO.id()) {
            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
        }
        
        boolean isFavorite = userService.isFavorite(userDTO, id);
        if(isFavorite){
            userService.removeFromFavorite(id, userDTO);
        }else{
            userService.addToFavorite(id, userDTO);
        }

        Page<ProductDTO> productsPage = productService.getFavoriteProducts(userDTO, page, size);
        return ResponseEntity.ok(productsPage);
    }

    @Operation (summary= "Retrieve last purchases of the user")
    @GetMapping("/purchases")
    public ResponseEntity<List<ProductDTO>> getLastPurchases(HttpServletRequest request) {

        Principal principal = request.getUserPrincipal();
        if (principal == null) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        UserDTO userDTO = userService.findByMailExtendedInfo(principal.getName()).get();

        List<ProductDTO> products = productService.getLastPurchases(userDTO);
        return ResponseEntity.ok(products);
    }

    @Operation (summary= "Retrieve last sales of the user")
    @GetMapping("/sales")
    public ResponseEntity<List<ProductDTO>> getLastSales(HttpServletRequest request) {

        Principal principal = request.getUserPrincipal();
        if (principal == null) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        UserDTO userDTO = userService.findByMailExtendedInfo(principal.getName()).get();

        List<ProductDTO> products = productService.getLastSales(userDTO);
        return ResponseEntity.ok(products);

    }

    @Operation (summary= "Add image to the product by the ID of the products")
    @PostMapping("/{id}/images")
    public ResponseEntity<Void> addImage(@PathVariable long id, @RequestParam MultipartFile image,
            HttpServletRequest request) throws IOException {
        Principal principal = request.getUserPrincipal();
        if (principal == null) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        UserBasicDTO userDTO = userService.findByMail(principal.getName()).get();
        Optional<ProductDTO> optionalProduct = productService.findById(id);
        if (optionalProduct.isEmpty()) {
            throw new NoSuchElementException();
        }

        ProductDTO productDTO = optionalProduct.get();
        if (productDTO.owner().id() != userDTO.id()) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        productService.addImageEditing(productDTO, image);
        return ResponseEntity.ok().build();

    }

    @Operation (summary= "Delete image of a product by image ID")
    @DeleteMapping("/{productId}/images/{imageId}")
    public ResponseEntity<Void> removeImage(@PathVariable long productId, @PathVariable long imageId,
            HttpServletRequest request) {
        Principal principal = request.getUserPrincipal();
        if (principal == null) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        UserBasicDTO userDTO = userService.findByMail(principal.getName()).get();
        Optional<ProductDTO> optionalProduct = productService.findById(productId);
        if (optionalProduct.isEmpty()) {
            throw new NoSuchElementException();
        }

        ProductDTO productDTO = optionalProduct.get();
        if (productDTO.owner().id() != userDTO.id()) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        if (!productService.imageBelongsToProduct(productId, imageId)) {
            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
        }

        productService.removeImage(productId, imageId);
        return ResponseEntity.ok().build();

    }

}